wxSpinCtrl::SetIncrement is supposed to take an int, not a double
git-svn-id: https://svn.wxwidgets.org/svn/wx/wxWidgets/trunk@62583 c3d73ce0-8a6f-49c7-b76d-6d57e0e08775

+++ include/wx/gtk/spinctrl.h
@@ -126,7 +126,7 @@ public:
     void SetValue(const wxString& value)    { wxSpinCtrlGTKBase::SetValue(value); } // visibility problem w/ gcc
     void SetValue( int value )              { DoSetValue(value); }
     void SetRange( int minVal, int maxVal ) { DoSetRange(minVal, maxVal); }
-    void SetIncrement( double inc )         { DoSetIncrement(inc); }
+    void SetIncrement(int inc) { DoSetIncrement(inc); }
 
     DECLARE_DYNAMIC_CLASS(wxSpinCtrl)
 };
